Fixed memory leak in repl threads.
The memory leak happened if we had a bunch of different "headers" (i.e.,
different sets of imports for different theorems rather than a
ubiquitous "import Mathlib" or whatever).

Basically, we would handle different headers by opening a new
tcp connection to an existing REPL process, initializing it with a
particular header, and then caching that (header, connection) pair, so
now if we got a new input theorem with the same header, we would simply
reuse that connection. To prevent too much memory being used, we would
cap the cache at K open sockets (i.e., K different environments per
worker).

Assuming this idea works, I had a few bugs in the implementation.

1. First of all, because I was using @lru_cache, it was also trying to
   hash the `self` object as a cache key, which honestly might work, but
I don't know because `self` was mutable and was changing. I don't know
how the hash function worked for that, so I just rolled my own with a
custom key_fn parameter to control the key.

2. Secondly, when something was evicted from the cache, we wouldn't
   actually close the socket connection - it would just have no
references. I don't know if the garbage collector would take care of
this but now we don't have to worry about that, since I added `del_fn`

Those are fixed now.

However, the core problem was actually in LeanREPL, not this project.
Essentially, every time we spawned a new environment (i.e., every time
we open a connection with a new header), **that environment would never
be deleted, even after the connection closed**. I added some code which
evicted an environment, but I'm not fluent enough with the Lean garbage
collection algorithm to know why this wasn't working. To evict, I was
popping the env from the cmdStates array in the StateT monad. I printed
out the size of the cmdStates array to make sure it was being deleted,
and it was. My only guess is that Lean doesn't actually deallocate the
memory, it just decreases a 'size' object or something. Not sure how to
force it.

So I couldn't figure out how to solve this in the LeanREPL side.
Instead, I fixed it in blv by adjusting how we handle new headers. Now,
instead of opening a new connection to the same REPL, I just launch a
new REPL process, and kill the process on eviction from the cache. I
realized that opening a REPL process is actually lightweight since we're
not loading any unnecessary libraries on launch, just starting the TCP
server. So without any real performance loss, we solve the issue,
because the REPL memory is deallocated when the process is killed, thus
we only ever use memory for in-cache processes.

The end.